\name{CombMSC-package}
\alias{CombMSC-package}
\alias{CombMSC}
\docType{package}
\title{
Combined Model Selection Criteria
}
\description{
Functions to compute optimal convex combinations of model selection criteria,
as well as utility functions to construct model lists and priors, and several
common model selection criteria.
}
\details{
\tabular{ll}{
Package: \tab CombMSC\cr
Type: \tab Package\cr
Version: \tab 1.4\cr
Date: \tab 2007-11-27\cr
License: \tab GPL v3\cr
}
The most important function is \code{\link{TMC}}.  The other user-level functions are mainly subsidiary,
and are used, for example, to construct objects to be passed as arguments to \code{\link{TMC}}.  
}
\author{
Andrew K. Smith

Maintainer: Andrew K. Smith <andrewsmith81@gmail.com>

}


\keyword{ package }
