\name{gen.Data}
\alias{gen.Data}
\alias{gen.Data.glmFormula}
\alias{gen.Data.lmFormula}
\alias{gen.Data.tsm}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{ Simulate Data Sets }
\description{
 Simulates data sets from a given regression or time series model.  For internal use
 in TMC.
}
\keyword{datagen}