\name{num.Terms}
\alias{num.Terms}
\alias{num.Terms.formula}
\alias{num.Terms.tsm}
\title{ Number of terms (parameters) in a model }
\description{
 For a regression formula or a numeric vector specifying a time series model,
 calculates the number of parameters needed to specify the model.  Mainly used
 to construct weight vectors based on the complexity of the models.
}
\author{Andrew K. Smith}
\keyword{models}