\name{print.summary.msc}
\alias{print.summary.msc}
\alias{summary.msc}
\title{ Default print method for a summary of an msc object}
\description{
  Prints the most relevant information from an msc object.  Gives the
  values of all functions in the summary.Functions component of the object
  evaluated at all of the corners (convex combinations with all weights equal to 0 or 1)
  , and then gives the top 5 convex combinations for each summary function, along with
  the values of all summary functions corresponding to those combinations.
}
\author{Andrew K. Smith}
\keyword{print}