\name{sgnf}
\alias{sgnf}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{ Significance of an msc object }
\description{
  A convenience function which calculates, for each summary function of the msc object,
  the difference between the best pure MSC and the best combined MSC (i.e., how much better
  we can make each summary function by considering combined MSC instead of only the pure MSCs.)
  The results may be useful if one is interested in testing the hypothesis that pure MSCs are as
  good as any convex combination of MSCs.
}


\author{Andrew K. Smith}
\keyword{print}